Refactor income and expense models to use BigDecimal for amounts; add expense adjustment handling and validation exceptions

+++ src/main/java/com/expensetracker/dto/ExpenseRequest.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,7 +14,7 @@ public class ExpenseRequest {
     private Integer expensesId; // optional for updates
     private String userId;
     private String expenseName;
-    private Double expenseAmount;
+    private BigDecimal expenseAmount;
     private Integer userExpenseCategoryId;
     private LocalDate expenseDate;
 }
